refactor(mobile): extract useNoteFormattingToolbar shared hook
The note composer (app/note/new.tsx) and the note editor (NoteDetailScreen)
had ~65 lines of identical editor-toolbar state + callbacks duplicated verbatim
(onSelectionChange, onFormat, onWikilink, onAIAction, hasSelection, selection,
aiLoading, wikilinkOpen). Any formatting/AI-action fix had to be made twice and
they would inevitably drift.

Move the shared logic into src/notes/useNoteFormattingToolbar(body, setBody);
both screens now consume it with an identical API (screens still own body/setBody
since NoteDetailScreen resets body on sync/refocus). Behaviour is unchanged.

Net -150 lines across the two screens. Also drops a now-stale AppleMessage
import in the apple-llm bridge (generate takes a prompt string, not messages).

type-check + lint clean (only pre-existing set-state-in-effect warnings).
