+++ mobile/src/screens/NoteDetailScreen.tsx
@@ -7,8 +7,6 @@ import {
   StyleSheet,
   View,
   Alert,
-  type NativeSyntheticEvent,
-  type TextInputSelectionChangeEventData,
 } from "react-native";
 import { Pin } from "lucide-react-native";
 import { KeyboardAwareScrollView, KeyboardStickyView } from "react-native-keyboard-controller";
@@ -20,9 +18,7 @@ import { TagPickerSheet } from "@/components/TagPickerSheet";
 import { NoteEditorToolbar } from "@/components/NoteEditorToolbar";
 import { WikilinkPickerSheet } from "@/components/WikilinkPickerSheet";
 import { ICON_CHECK, ICON_CLOSE, ICON_EDIT, ICON_MORE, ICON_PIN, ICON_UNPIN, ICON_TAG, ICON_DELETE } from "@/components/toolbar-icons";
-import { applyFormat, insertWikilink, type FormatAction, type Selection } from "@cairn/shared/notes/format";
-import { buildAIActionPrompt, type AITextAction } from "@cairn/shared/notes/ai-actions";
-import { runTextAction } from "@/chat/agent";
+import { useNoteFormattingToolbar } from "@/notes/useNoteFormattingToolbar";
 import { useDataChanged } from "@/sync/useSyncStatus";
 import { useTheme, TAB_BAR_BASE, type as typeScale, type Theme } from "@/theme";
 
@@ -53,10 +49,7 @@ export function NoteDetailScreen({ nested = false }: { nested?: boolean }) {
 
   // Editing toolbar state.
   const bodyRef = useRef<TextInput>(null);
-  const [selection, setSelection] = useState<Selection>({ start: 0, end: 0 });
-  const selectionRef = useRef<Selection>({ start: 0, end: 0 });
-  const [aiLoading, setAiLoading] = useState(false);
-  const [wikilinkOpen, setWikilinkOpen] = useState(false);
+  const fmt = useNoteFormattingToolbar(body, setBody);
 
   const styles = useMemo(() => makeStyles(t), [t]);
 
@@ -73,75 +66,6 @@ export function NoteDetailScreen({ nested = false }: { nested?: boolean }) {
   }, [editing, id]);
   useFocusEffect(useCallback(() => reload(), [reload]));
   useDataChanged(reload);
-
-  const onSelectionChange = useCallback(
-    (e: NativeSyntheticEvent<TextInputSelectionChangeEventData>) => {
-      selectionRef.current = e.nativeEvent.selection;
-      setSelection(e.nativeEvent.selection);
-    },
-    [],
-  );
-
-  const hasSelection = selection.end > selection.start;
-
-  // Apply a formatting action to the body at the current selection, then restore
-  // the (shifted) selection so the caret lands sensibly.
-  const onFormat = useCallback(
-    (action: FormatAction) => {
-      if (action === "wikilink") {
-        setWikilinkOpen(true);
-        return;
-      }
-      const res = applyFormat(body, selectionRef.current, action);
-      if (!res) return;
-      setBody(res.text);
-      selectionRef.current = res.selection;
-      setSelection(res.selection);
-    },
-    [body],
-  );
-
-  const onWikilink = useCallback(
-    (noteTitle: string) => {
-      const res = insertWikilink(body, selectionRef.current, noteTitle);
-      setBody(res.text);
-      selectionRef.current = res.selection;
-      setSelection(res.selection);
-      setWikilinkOpen(false);
-    },
-    [body],
-  );
-
-  // Run an AI text action over the current selection and replace it with the
-  // model's reply (online only, via Rork).
-  const onAIAction = useCallback(
-    async (action: AITextAction, customPrompt?: string) => {
-      const sel = selectionRef.current;
-      const selected = body.slice(sel.start, sel.end);
-      if (!selected) return;
-      setAiLoading(true);
-      try {
-        const prompt = buildAIActionPrompt(action, selected, customPrompt);
-        const reply = await runTextAction(prompt);
-        if (!reply) return;
-        const next = body.slice(0, sel.start) + reply + body.slice(sel.end);
-        const newSel = { start: sel.start, end: sel.start + reply.length };
-        setBody(next);
-        selectionRef.current = newSel;
-        setSelection(newSel);
-      } catch (e) {
-        Alert.alert(
-          "AI action failed",
-          e instanceof Error && /network|fetch|connect|\(5\d\d\)/i.test(e.message)
-            ? "This needs a connection. Reconnect and try again."
-            : "Something went wrong. Try again.",
-        );
-      } finally {
-        setAiLoading(false);
-      }
-    },
-    [body],
-  );
 
   const isPinned = !!note?.is_pinned;
 
@@ -266,8 +190,8 @@ export function NoteDetailScreen({ nested = false }: { nested?: boolean }) {
               style={styles.bodyInput}
               value={body}
               onChangeText={setBody}
-              selection={selection}
-              onSelectionChange={onSelectionChange}
+              selection={fmt.selection}
+              onSelectionChange={fmt.onSelectionChange}
               placeholder="Write in Markdown…"
               placeholderTextColor={t.textTertiary}
               multiline
@@ -276,11 +200,11 @@ export function NoteDetailScreen({ nested = false }: { nested?: boolean }) {
           </KeyboardAwareScrollView>
           <KeyboardStickyView>
             <NoteEditorToolbar
-              onFormat={onFormat}
-              onAction={onAIAction}
-              hasSelection={hasSelection}
+              onFormat={fmt.onFormat}
+              onAction={fmt.onAIAction}
+              hasSelection={fmt.hasSelection}
               aiEnabled
-              loading={aiLoading}
+              loading={fmt.aiLoading}
               onDismiss={() => bodyRef.current?.blur()}
               bottomInset={insets.bottom + (nested ? TAB_BAR_BASE : 0)}
             />
@@ -308,9 +232,9 @@ export function NoteDetailScreen({ nested = false }: { nested?: boolean }) {
       />
 
       <WikilinkPickerSheet
-        visible={wikilinkOpen}
-        onSelect={onWikilink}
-        onClose={() => setWikilinkOpen(false)}
+        visible={fmt.wikilinkOpen}
+        onSelect={fmt.onWikilink}
+        onClose={fmt.closeWikilink}
       />
     </View>
   );
